requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.6
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
